ref: move tasks to a separate component
