+++ src/components/CreateTodo/style.module.scss
@@ -36,57 +36,6 @@
       }
     }
   }
-  .containerTasks {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    height: 500px;
-    width: 1060px;
-    border: 2px solid #bababa;
-    border-radius: 20px;
-    box-shadow: 0 0 10px 1px #6a55ef;
-    background: linear-gradient(125deg, #a8c3fa, #a89bff);
-    .caption {
-      height: 250px;
-      display: flex;
-      align-items: center;
-      span {
-        font-size: 28px;
-        font-weight: 600;
-        color: white;
-      }
-    }
-    .titleTasks {
-      width: 100%;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      padding: 10px 0;
-      border-radius: 20px 20px 0 0;
-      span {
-        font-size: 34px;
-        text-align: center;
-        letter-spacing: 30px;
-        margin: 0 -30px 0 15px;
-        font-weight: 800;
-        color: white;
-        font-family: "Courier New", Courier, monospace;
-      }
-    }
-    .tasks {
-      width: 92.5%;
-      height: auto;
-      display: flex;
-      justify-content: flex-start;
-      flex-wrap: wrap;
-      overflow: hidden;
-      overflow-y: scroll;
-      padding: 20px 0;
-      &::-webkit-scrollbar {
-        width: 0px;
-      }
-    }
-  }
 }
 @media (max-width: 1200px) {
   .container {
@@ -101,9 +50,6 @@
 }
 @media (max-width: 768px) {
   .container {
-    .containerTasks {
-      width: 70%;
-    }
     .containerCard {
       width: 40%;
       height: 260px;
@@ -111,19 +57,6 @@
       .heading {
         font-size: 18px;
         font-weight: 600;
-      }
-    }
-    .containerTasks {
-      .tasks {
-        justify-content: center;
-      }
-      .titleTasks {
-        span {
-          font-size: 28px;
-          letter-spacing: 25px;
-          margin: 0 -25px 0 25px;
-          font-weight: 700;
-        }
       }
     }
   }
@@ -137,19 +70,6 @@
       .heading {
         font-size: 16px;
         font-weight: 600;
-      }
-    }
-    .containerTasks {
-      .tasks {
-        justify-content: center;
-      }
-      .titleTasks {
-        span {
-          font-size: 25px;
-          letter-spacing: 15px;
-          margin: 0 -15px 0 15px;
-          font-weight: 600;
-        }
       }
     }
   }
