+++ src/components/NavigationMenu/NavigationMenu.tsx
@@ -1,18 +1,16 @@
-import React, { FunctionComponent, useState, useEffect } from 'react';
-import { useLocation, NavLink } from 'react-router-dom';
+import React, { FunctionComponent } from 'react';
+import { NavLink } from 'react-router-dom';
+
+import { useSelector } from 'react-redux';
+
+import { RootState } from '../../types/types';
 
 import logo from '../../images/logo.png';
 
 import styles from './style.module.scss';
 
 export const NavigationMenu: FunctionComponent = (): JSX.Element => {
-  const [currentPath, setCurrentPath] = useState<string>('');
-
-  const location = useLocation();
-
-  useEffect(() => {
-    setCurrentPath(() => location.pathname);
-  }, [location]);
+  const amountTodo = useSelector((state: RootState) => state?.todosReducers?.length);
 
   return (
     <div className={styles.container}>
@@ -21,11 +19,9 @@ export const NavigationMenu: FunctionComponent = (): JSX.Element => {
           <img src={logo} alt="logo" />
           <span className={styles.logo}>Task Managment</span>
         </NavLink>
-        <div className={styles.navPath}>
-          <span className={styles.titlePath}>
-            {currentPath === '/' ? 'Home' : 'Edit'}
-          </span>
-        </div>
+        <NavLink to="/todos" className={styles.navLink}>
+          <span className={styles.logo}>{`ToDo's (${amountTodo})`}</span>
+        </NavLink>
       </div>
     </div>
   );
