+++ src/components/Todos/style.module.scss
@@ -1,0 +1,102 @@
+.container {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  align-items: center;
+  .containerTasks {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    height: 600px;
+    width: 1060px;
+    border: 2px solid #bababa;
+    border-radius: 20px;
+    box-shadow: 0 0 10px 1px #6a55ef;
+    background: linear-gradient(125deg, #a8c3fa, #a89bff);
+    margin: 30px 0 0 0;
+    .caption {
+      height: 250px;
+      display: flex;
+      align-items: center;
+      span {
+        font-size: 28px;
+        font-weight: 600;
+        color: white;
+      }
+    }
+    .titleTasks {
+      width: 100%;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      padding: 10px 0;
+      border-radius: 20px 20px 0 0;
+      span {
+        font-size: 34px;
+        text-align: center;
+        letter-spacing: 30px;
+        margin: 0 -30px 0 15px;
+        font-weight: 800;
+        color: white;
+        font-family: "Courier New", Courier, monospace;
+      }
+    }
+    .tasks {
+      width: 92.5%;
+      height: auto;
+      display: flex;
+      justify-content: flex-start;
+      flex-wrap: wrap;
+      overflow: hidden;
+      overflow-y: scroll;
+      padding: 20px 0;
+      &::-webkit-scrollbar {
+        width: 0px;
+      }
+    }
+  }
+}
+@media (max-width: 1200px) {
+  .container {
+    .containerTasks {
+      width: 70%;
+    }
+  }
+}
+@media (max-width: 768px) {
+  .container {
+    .containerTasks {
+      width: 70%;
+      .tasks {
+        justify-content: center;
+      }
+
+      .titleTasks {
+        span {
+          font-size: 28px;
+          letter-spacing: 25px;
+          margin: 0 -25px 0 25px;
+          font-weight: 700;
+        }
+      }
+    }
+  }
+}
+@media (max-width: 534px) {
+  .container {
+    .containerTasks {
+      .tasks {
+        justify-content: center;
+      }
+      .titleTasks {
+        span {
+          font-size: 25px;
+          letter-spacing: 15px;
+          margin: 0 -15px 0 15px;
+          font-weight: 600;
+        }
+      }
+    }
+  }
+}
